Small changes, move old version into folder
