+++ player.py
@@ -19,10 +19,6 @@ class Player:
         pass
 
     @abstractmethod
-    def pick_up_card(self, suit: str, flipped_c: Card):
-        pass
-
-    @abstractmethod
     def choose_call_suit(self, suit: str, flipped_c: Card):
         pass   
 
@@ -57,3 +53,12 @@ class Player:
     @abstractmethod
     def choose_card(self):
         pass   
+
+class Team:
+    # Teams are made up of 2 players. In real Euchre, teammates sit across the table from each other. So odd players
+    # make up team1, and even numbered players make up team2
+    def __init__(self, player_a, player_b, points, tricks):
+        self.player_a = player_a
+        self.player_b = player_b
+        self.points = points
+        self.tricks = tricks
